small fixes

+++ cpp07/ex01/iter.hpp
@@ -3,10 +3,10 @@
 # include <iostream>
 
 template <typename T>
-void iter(T *arr, size_t length, void (*foo)(T &)) {
+void iter(T *arr, int length, void (*foo)(T &)) {
     if (!arr || !foo)
         return ;
-    for (size_t i = 0; i < length; i++){
+    for (int i = 0; i < length; i++){
         foo(arr[i]);
     }
 }
