+++ cpp07/ex01/main.cpp
@@ -1,5 +1,6 @@
 # include "iter.hpp"
 # include <iostream>
+# include <strings.h>
 # include <cctype>
 
 void toUpper(char &e)
@@ -14,28 +15,28 @@ int main ()
     std::string strArray[] = {"str1", "str2", "str3"};
 
 
-    for (size_t i=0; i < 4; i++){
+    for (int i=0; i < 4; i++){
         std::cout << charArray[i] << std::endl;
     }
     ::iter(charArray, 4, toUpper);
-    for (size_t i=0; i < 4; i++){
+    for (int i=0; i < 4; i++){
         std::cout << charArray[i] << std::endl;
     }
 
-    for (size_t i=0; i < 4; i++){
+    for (int i=0; i < 4; i++){
         std::cout << intArray[i] << std::endl;
     }
     ::iter(charArray, 4, toUpper);
-    for (size_t i=0; i < 4; i++){
+    for (int i=0; i < 4; i++){
         std::cout << intArray[i] << std::endl;
     }
 
     // Not Working
-    // for (size_t i=0; i < 3; i++){ 
+    // for (int i=0; i < 3; i++){ 
     //     std::cout << strArray[i] << std::endl;
     // }
     // ::iter(strArray, 3, toUpper);
-    // for (size_t i=0; i < 3; i++){
+    // for (int i=0; i < 3; i++){
     //     std::cout << strArray[i] << std::endl;
     // }
